+++ src/utils/simplify.js
@@ -5,12 +5,14 @@ const simplify = username =>
   R.pipe(
     R.path(['payload', 'references', 'Post']),
     R.values,
-    R.project(['title', 'uniqueSlug', 'updatedAt']),
+    R.project(['title', 'uniqueSlug', 'firstPublishedAt', 'content']),
     R.map(e => ({
       title: e.title,
-      updatedAt: e.updatedAt,
+      firstPublishedAt: e.firstPublishedAt,
       url: `${BASE_URL}/@${username}/${e.uniqueSlug}`,
+      subtitle: e.content.subtitle,
     })),
+    R.sort(R.descend(R.prop('firstPublishedAt'))),
   );
 
 module.exports = simplify;